requirements.txt
pandas
datasets
openai
python-dotenv
torch